type check recursive types

+++ src/main/java/org/yinwang/yin/value/FloatType.java
@@ -3,6 +3,6 @@ package org.yinwang.yin.value;
 public class FloatType extends Value {
 
     public String toString() {
-        return "Int";
+        return "Float";
     }
 }
